refactor: Lotto 관련 상수들 전역에서 사용하는 패키지로 분리
