+++ src/main/java/lotto/constant/LottoConstant.java
@@ -1,0 +1,7 @@
+package lotto.constant;
+
+public class LottoConstant {
+    public static final int MIN_NUMBER = 1;
+    public static final int MAX_NUMBER = 45;
+    public static final int SIZE = 6;
+}
