+++ src/main/java/lotto/model/Lotto.java
@@ -1,12 +1,9 @@
 package lotto.model;
 
 import java.util.List;
+import lotto.constant.LottoConstant;
 
 public class Lotto {
-    private static final int LOTTO_LENGTH = 6;
-    private static final int LOTTO_LEFT_BOUND = 1;
-    private static final int LOTTO_RIGHT_BOUND = 45;
-
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
@@ -21,13 +18,13 @@ public class Lotto {
     }
 
     private void validateLengthCheck(List<Integer> numbers) {
-        if (numbers.size() != LOTTO_LENGTH) {
-            throw new IllegalArgumentException(String.format("로또 번호의 갯수 %d개를 만족해야합니다.", LOTTO_LENGTH));
+        if (numbers.size() != LottoConstant.SIZE) {
+            throw new IllegalArgumentException(String.format("로또 번호의 갯수 %d개를 만족해야합니다.", LottoConstant.SIZE));
         }
     }
 
     private void validateDuplicateNumber(List<Integer> numbers) {
-        if (getDistinctCount(numbers) != LOTTO_LENGTH) {
+        if (getDistinctCount(numbers) != LottoConstant.SIZE) {
             throw new IllegalArgumentException("숫자는 중복될 수 없습니다.");
         }
     }
@@ -37,11 +34,11 @@ public class Lotto {
     }
 
     private void validateProperNumber(Integer number) {
-        if (number < LOTTO_LEFT_BOUND || number > LOTTO_LEFT_BOUND) {
+        if (number < LottoConstant.MIN_NUMBER || number > LottoConstant.MAX_NUMBER) {
             throw new IllegalArgumentException(String.format(
                     "숫자는 %d에서 %d사이여야합니다."
-                    , LOTTO_LEFT_BOUND
-                    , LOTTO_RIGHT_BOUND
+                    , LottoConstant.MIN_NUMBER
+                    , LottoConstant.MAX_NUMBER
                 )
             );
         }
