+++ src/test/java/lotto/model/LottoTest.java
@@ -1,9 +1,8 @@
-package lotto;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
